requirements.txt
# This is the swatchable.io python requirements
Pillow